Adding a new function for to_talk methods.
Each will be passed a single parameter: a boolean. If it is true, we are
synchronizing at the moment and it is advisable to pass all relevant data.
If it is false, we are in normal operation and thus, to conserve on network
usage, only pass the data necessary to identify/deserialize the class.